fix(sigsession): 修复工作模式切换时通道使能不同步问题
在切换工作模式时同步所有通道的使能状态，解决上游libsigrok演示驱动在非对应模式下仍发送数据导致视图为空的问题。同时修复headerData中的类型转换问题，使用qlonglong替代强制转换避免潜在的类型不匹配。

+++ PXView/pv/view/decodermodel.cpp
@@ -115,7 +115,7 @@ QVariant DecoderModel::headerData(int section,
 
     if (orientation == Qt::Vertical) {
         if (_visible_start_row >= 0)
-            return (int64_t)section + _visible_start_row;
+            return qlonglong((int64_t)section + _visible_start_row);
         return section;
     }
